피곤스

+++ Project/src/main/java/com/mvc/grade/GradeDTO.java
@@ -16,5 +16,5 @@ public class GradeDTO {
 	
 	private Long listCode;
 	
-	private Long grade;
+	private Double grade;
 }
